Petites modif. -Le programme piano_audio_to_piano_note tourne mais le calcul des transformés de fourier est trop long, une piste -> convertir les float en short pour réduir le temps de calcul

+++ src/lect_audio.c
@@ -83,7 +83,7 @@ void rempli_tab(FILE *wav, float* tab_temps, float* tab_amplitudes){
 
     //Remplissage du tableau
     int i=0;
-    short value=0;
+    float value=0;
     while(fread(&value,(header.bits_per_sample)/8,1,wav)){
         //lecture des echantillons et enregistrement dans le tableau (en dB)
         tab_amplitudes[i] = 20*log10(value);
